pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "py-opendisplay"
version = "0.1.0"
description = "Python library for OpenDisplay BLE e-paper displays"
readme = "README.md"
license = "MIT"
authors = [{name = "g4bri3lDev", email = "admin@g4bri3l.de"}]
requires-python = ">=3.11"

dependencies = [
    "bleak>=1.0.1",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://opendisplay.org"
Repository = "https://github.com/OpenDisplay-org/py-opendisplay"

[tool.hatch.build.targets.wheel]
packages = ["src/opendisplay"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"